--- lib/features/main/BookDetail.dart
@@ -1,134 +1,0 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_rating_bar/flutter_rating_bar.dart';
-
-class BookDetail extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return SafeArea(
-      child: Material(
-        color: Colors.indigo,
-        child: CustomScrollView(
-          slivers: [
-            SliverPersistentHeader(
-              delegate: MySliverAppBar(expandedHeight: 150),
-              pinned: true,
-            ),
-            SliverList(
-              delegate: SliverChildListDelegate([
-                Container(
-                  width: double.infinity,
-                  decoration: BoxDecoration(
-                      borderRadius: BorderRadius.only(
-                          topLeft: Radius.circular(5),
-                          topRight: Radius.circular(5)),
-                      color: Colors.white),
-                  child: Column(
-                    children: <Widget>[
-                      SizedBox(
-                        height: 200,
-                      ),
-                      Text("Be a Copywrite",
-                          style: TextStyle(fontSize: 24, color: Colors.black)),
-                      SizedBox(
-                        height: 4,
-                      ),
-                      Text(
-                        "Minor Developer",
-                        style: TextStyle(fontSize: 14, color: Colors.grey),
-                      ),
-                      RatingBar(
-                        initialRating: 3,
-                        minRating: 1,
-                        direction: Axis.horizontal,
-                        allowHalfRating: true,
-                        itemCount: 5,
-                        itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                        itemBuilder: (context, _) => Icon(
-                          Icons.star,
-                          color: Colors.amber,
-                        ),
-                        onRatingUpdate: (rating) {
-                          print(rating);
-                        },
-                      )
-                    ],
-                  ),
-                )
-              ]),
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class MySliverAppBar extends SliverPersistentHeaderDelegate {
-  final double expandedHeight;
-
-  MySliverAppBar({@required this.expandedHeight});
-
-  @override
-  Widget build(
-      BuildContext context, double shrinkOffset, bool overlapsContent) {
-    return Stack(
-      fit: StackFit.expand,
-      overflow: Overflow.visible,
-      children: [
-        Container(
-          color: Colors.indigo,
-        ),
-        Center(
-          child: Opacity(
-            opacity: shrinkOffset / expandedHeight,
-            child: Text(
-              "Book Details",
-              style: TextStyle(
-                color: Colors.white,
-                fontWeight: FontWeight.w700,
-                fontSize: 23,
-              ),
-            ),
-          ),
-        ),
-        Positioned(
-          top: expandedHeight * (0.6) - shrinkOffset,
-          left: MediaQuery.of(context).size.width / 4,
-          child: Opacity(
-            opacity: (1 - shrinkOffset / expandedHeight),
-            child: logoBook(context),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget logoBook(BuildContext context) {
-    return Container(
-      width: MediaQuery.of(context).size.width / 2,
-      height: 240,
-      child: Card(
-          semanticContainer: true,
-          elevation: 10,
-          clipBehavior: Clip.antiAliasWithSaveLayer,
-          shape: RoundedRectangleBorder(
-            side: BorderSide(color: Colors.white, width: 1),
-            borderRadius: BorderRadius.circular(5),
-          ),
-          child: Image.network(
-            'https://i.picsum.photos/id/9/160/200.jpg',
-            fit: BoxFit.fill,
-          )),
-    );
-  }
-
-  @override
-  double get maxExtent => expandedHeight;
-
-  @override
-  double get minExtent => kToolbarHeight;
-
-  @override
-  bool shouldRebuild(SliverPersistentHeaderDelegate oldDelegate) => true;
-}
